dr-saintvision/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dr-saintvision"
version = "1.0.0"
description = "Multi-Agent AI Debate System for Enhanced Reasoning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "DR-Saintvision Team"}
]
keywords = [
    "ai",
    "llm",
    "multi-agent",
    "debate",
    "reasoning",
    "rag",
    "mistral",
    "llama",
    "qwen"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.3",
    "sentencepiece",
    "protobuf",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "duckduckgo-search>=4.1.0",
    "beautifulsoup4>=4.12.2",
    "fastapi>=0.108.0",
    "uvicorn>=0.25.0",
    "gradio>=4.13.0",
    "sqlalchemy>=2.0.23",
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
dr-saintvision = "main:main"

[project.urls]
Homepage = "https://github.com/dr-saintvision/dr-saintvision"
Documentation = "https://github.com/dr-saintvision/dr-saintvision#readme"
Repository = "https://github.com/dr-saintvision/dr-saintvision"

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "frontend*", "backend*", "utils*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=html"
